\name{modopt.matlab-package}
\alias{modopt.matlab}

\docType{package}

\title{Modeling Optimization Problems}

\description{MatLab(R)-style Modeling of Optimization Problems with R}

\details{
\tabular{ll}{
Package: \tab modopt.matlab\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2014-04-09\cr
License: \tab GPL-2\cr
}
}

\author{
Ronald Hochreiter \email{ron@hochreiter.net}

Maintainer: Ronald Hochreiter \email{ron@hochreiter.net}
}

\references{\url{http://www.modopt.com/}}

\keyword{ package }

\examples{}