% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modopt.matlab.R
\docType{package}
\name{modopt.matlab-package}
\alias{modopt.matlab}
\alias{modopt.matlab-package}
\title{MatLab(R)-style Optimization Modeling in R using ROI}
\description{
\references{\url{http://www.finance-r.com/}}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.finance-r.com/}
}

}
\author{
Ronald Hochreiter, \email{ron@hochreiter.net}
}
